* Fixed a few minor warnings like the following:
cipher_funcs.c:85: warning: ignoring return value of ‘fread’, declared with attribute warn_unused_result

A few of these were in code in the lib/ directory, and required adding a
new error code 'FKO_ERROR_FILESYSTEM_OPERATION' and associated error
string 'Read/write bytes mismatch'.




git-svn-id: file:///home/mbr/svn/fwknop/trunk@195 510a4753-2344-4c79-9c09-4d669213fbeb

+++ lib/fko_encode.c
@@ -110,7 +110,7 @@ fko_encode_spa_data(fko_ctx_t ctx)
     /* Add the timestamp.
     */
     offset = strlen(tbuf);
-    sprintf(((char*)tbuf+offset), ":%u:", ctx->timestamp);
+    sprintf(((char*)tbuf+offset), ":%u:", (unsigned int) ctx->timestamp);
 
     /* Add the version string.
     */
